requirements.txt
# This file may be used to create an environment using:
# $ conda create --name <env> --file <this file>
# platform: osx-arm64
annotated-types=0.6.0=py312hca03da5_0
appnope=0.1.4=pyhd8ed1ab_0
asttokens=2.4.1=pyhd8ed1ab_0
blas=1.0=openblas
bottleneck=1.3.7=py312ha86b861_0
brotli-python=1.0.9=py312h313beb8_8
bzip2=1.0.8=h80987f9_6
ca-certificates=2024.8.30=hf0a4a13_0
catalogue=2.0.10=py312hca03da5_0
certifi=2024.8.30=py312hca03da5_0
charset-normalizer=3.3.2=pyhd3eb1b0_0
click=8.1.7=py312hca03da5_0
cloudpathlib=0.16.0=py312hca03da5_1
colorama=0.4.6=py312hca03da5_0
comm=0.2.2=pyhd8ed1ab_0
confection=0.1.4=py312h989b03a_0
cymem=2.0.6=py312h313beb8_0
cython-blis=0.7.11=py312ha86b861_0
debugpy=1.6.7=py312h313beb8_0
decorator=5.1.1=pyhd8ed1ab_0
exceptiongroup=1.2.2=pyhd8ed1ab_0
executing=2.1.0=pyhd8ed1ab_0
expat=2.6.3=h313beb8_0
filelock=3.13.1=py312hca03da5_0
freetype=2.12.1=h1192e45_0
fsspec=2024.6.1=py312hca03da5_0
idna=3.7=py312hca03da5_0
importlib-metadata=8.5.0=pyha770c72_0
ipykernel=6.29.5=pyh57ce528_0
ipython=8.27.0=pyh707e725_0
jedi=0.19.1=pyhd8ed1ab_0
jinja2=3.1.4=py312hca03da5_0
jpeg=9e=h80987f9_3
jupyter_client=8.6.3=pyhd8ed1ab_0
jupyter_core=5.7.2=pyh31011fe_1
langcodes=3.3.0=pyhd3eb1b0_0
lcms2=2.12=hba8e193_0
lerc=3.0=hc377ac9_0
libabseil=20240116.2=cxx17_h313beb8_0
libcxx=14.0.6=h848a8c0_0
libdeflate=1.17=h80987f9_1
libffi=3.4.4=hca03da5_1
libgfortran=5.0.0=11_3_0_hca03da5_28
libgfortran5=11.3.0=h009349e_28
libjpeg-turbo=2.0.0=h1a28f6b_0
libopenblas=0.3.21=h269037a_0
libpng=1.6.39=h80987f9_0
libprotobuf=4.25.3=h514c7bf_0
libsodium=1.0.18=h27ca646_1
libtiff=4.5.1=h313beb8_0
libwebp-base=1.3.2=h80987f9_0
llvm-openmp=14.0.6=hc6e5704_0
lz4-c=1.9.4=h313beb8_1
markdown-it-py=2.2.0=py312hca03da5_1
markupsafe=2.1.3=py312h80987f9_0
matplotlib-inline=0.1.7=pyhd8ed1ab_0
mdurl=0.1.0=py312hca03da5_0
mpmath=1.3.0=py312hca03da5_0
murmurhash=1.0.7=py312h313beb8_0
ncurses=6.4=h313beb8_0
nest-asyncio=1.6.0=pyhd8ed1ab_0
networkx=3.2.1=py312hca03da5_0
numexpr=2.8.7=py312h0f3ea24_0
numpy=1.26.4=py312h7f4fdc5_0
numpy-base=1.26.4=py312he047099_0
openjpeg=2.5.2=h54b8e55_0
openssl=3.3.2=h8359307_0
packaging=24.1=pyhd8ed1ab_0
pandas=2.2.2=py312hd77ebd4_0
parso=0.8.4=pyhd8ed1ab_0
pexpect=4.9.0=pyhd8ed1ab_0
pickleshare=0.7.5=py_1003
pillow=10.4.0=py312h80987f9_0
pip=24.2=py312hca03da5_0
platformdirs=4.3.6=pyhd8ed1ab_0
portalocker=2.10.1=pypi_0
preshed=3.0.6=py312h313beb8_0
prompt-toolkit=3.0.47=pyha770c72_0
psutil=5.9.0=py312h80987f9_0
ptyprocess=0.7.0=pyhd3deb0d_0
pure_eval=0.2.3=pyhd8ed1ab_0
pydantic=2.8.2=py312hca03da5_0
pydantic-core=2.20.1=py312hf0e4da2_0
pygments=2.18.0=pyhd8ed1ab_0
pysocks=1.7.1=py312hca03da5_0
python=3.12.4=h99e199e_1
python-dateutil=2.9.0post0=py312hca03da5_2
python-tzdata=2023.3=pyhd3eb1b0_0
pytorch=2.3.0=cpu_py312h6e465e2_1
pytz=2024.1=py312hca03da5_0
pyyaml=6.0.1=py312h80987f9_0
pyzmq=25.1.2=py312h313beb8_0
readline=8.2=h1a28f6b_0
requests=2.32.3=py312hca03da5_0
rich=13.7.1=py312hca03da5_0
setuptools=75.1.0=py312hca03da5_0
shellingham=1.5.0=py312hca03da5_0
six=1.16.0=pyhd3eb1b0_1
sleef=3.5.1=h80987f9_2
smart_open=5.2.1=py312hca03da5_0
spacy=3.7.2=py312h6d20652_0
spacy-legacy=3.0.12=py312hca03da5_0
spacy-loggers=1.0.4=py312hca03da5_0
sqlite=3.45.3=h80987f9_0
srsly=2.4.8=py312h313beb8_1
stack_data=0.6.2=pyhd8ed1ab_0
sympy=1.13.2=py312hca03da5_0
thinc=8.2.2=py312h6d20652_0
tk=8.6.14=h6ba3021_0
torchdata=0.8.0=py312
torchtext=0.18.0=py312
torchvision=0.18.1=cpu_py312hb6554cd_0
tornado=6.4.1=py312h80987f9_0
tqdm=4.66.5=py312h989b03a_0
traitlets=5.14.3=pyhd8ed1ab_0
typer=0.9.0=py312hca03da5_0
typing-extensions=4.11.0=py312hca03da5_0
typing_extensions=4.11.0=py312hca03da5_0
tzdata=2024a=h04d1e81_0
urllib3=2.2.2=py312hca03da5_0
wasabi=0.9.1=py312hca03da5_0
wcwidth=0.2.13=pyhd8ed1ab_0
weasel=0.3.4=py312hca03da5_0
wheel=0.44.0=py312hca03da5_0
xz=5.4.6=h80987f9_1
yaml=0.2.5=h1a28f6b_0
zeromq=4.3.5=h313beb8_0
zipp=3.20.2=pyhd8ed1ab_0
zlib=1.2.13=h18a0788_1
zstd=1.5.5=hd90d995_2
